back/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
audioread==3.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
exceptiongroup==1.2.0
faiss==1.7.4
faiss-cpu==1.7.4
fastapi==0.104.1
h11==0.14.0
idna==3.4
joblib==1.3.2
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
msgpack==1.0.7
nltk==3.8.1
numba==0.58.1
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1695290879892/work/dist/numpy-1.26.0-cp39-cp39-linux_x86_64.whl#sha256=2145d30cdd6d3c9c4a4be216bb158edeac64e1c9aba388e733918d65b9481735
packaging==23.2
pandas==2.1.3
platformdirs==4.0.0
pooch==1.8.0
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
Rtree==1.1.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sklearn==0.0.post11
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
starlette==0.27.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.24.0.post1
